+++ paper_admin/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "3.0.4"
+__version__ = "3.1.0"
 default_app_config = "paper_admin.apps.Config"