+++ tests/sortables/models.py
@@ -2,6 +2,7 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from mptt.models import MPTTModel, TreeForeignKey
 from solo.models import SingletonModel
+from tree_queries.models import TreeNode
 
 
 class Company(models.Model):
@@ -53,18 +54,31 @@ class CompanyIndustry(models.Model):
         return str(self.industry)
 
 
-class Category(MPTTModel):
+class MPTTTree(MPTTModel):
     parent = TreeForeignKey("self", null=True, blank=True, related_name="children", on_delete=models.CASCADE)
     title = models.CharField(_("title"), max_length=255)
-    order = models.PositiveIntegerField(_("order"), default=0, editable=False)
+    position = models.PositiveIntegerField(_("position"), default=0, editable=False)
 
     class Meta:
-        ordering = ["order"]
-        verbose_name = _("category")
-        verbose_name_plural = _("categories")
+        ordering = ["position"]
+        verbose_name = _("MPTT")
+        verbose_name_plural = _("MPTT")
 
     class MPTTMeta:
-        order_insertion_by = ["order"]
+        order_insertion_by = ["position"]
+
+    def __str__(self):
+        return self.title
+
+
+class DjangoTreeQueriesNode(TreeNode):
+    title = models.CharField(_("title"), max_length=255)
+    position = models.PositiveIntegerField(_("position"), default=0, editable=False)
+
+    class Meta:
+        ordering = ("position",)
+        verbose_name = _("Django Tree Queries")
+        verbose_name_plural = _("Django Tree Queries")
 
     def __str__(self):
         return self.title